Move Response and header types to own files
